ダウンロードツール/requirements.txt
playwright>=1.40.0
pandas>=2.0.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
pymupdf>=1.20.0
PySide6>=6.5.0
